requirements.txt
Flask==2.1.1
Flask-SQLAlchemy==2.5.1
Flask-Login==0.5.0
Flask-WTF==1.0.0
pycryptodome==3.10.1
werkzeug==2.0.3
sqlalchemy==1.4.49